fill_pageproject_aggregates: Fix query performance with composite index
When creating page aggregates we have a query that checks if an existing
aggregate record already exists for a page for the event's date. This
was fast for most collections, however on some colections it was
resulting in a full table scan causing a query that should only take
milliseconds to last 10 seconds or longer.

* Added a composite index for the fields 'full_date', 'collection_id',
  'project_name', and 'page_name' on the
  'aggregates_pageprojectaggregate' table.

* Changed the page aggregate query to remove the implicit 'ORDER BY'
  clause added by Django that was resulting in the new index not being
  used consistently.

Bug: T386399